Update README.md
The agile mindset is a set of values and principles extracted from the Agile Manifesto focused on trust, responsibility, ownership, continuous improvement, and continuous openness to learning and growing. A growing body of literature and anecdotal experiences in the gray literature suggest it is a necessary component for successful agile transformations in organizations. However, many organizations are unable to reap the full benefits of agile software engineering as they focus on agile practices only. Likewise, we postulate that the fully prepared graduate from a university software engineering program should exhibit an agile mindset, instead of merely gaining competence in agile practices. In this study, we investigate the agile mindset of university students in software engineering and working in an agile project environment as a part of a course. Multiple courses utilized a typical approach, teaching agile development competencies in a project-centric course. A third course extended this form of learning with additional critical inquiry activities to elevate internalization of agile principles and develop an agile mindset. A custom survey was employed and analyzed the results using standard descriptive and inferential statistics to investigate the outcomes.